mobile: re-organise file structure
