+++ mobile-app/src/utilities/functions/app-maker/checkIfActionTreeIsPopulated.ts
