+++ mobile-app/src/utilities/functions/app-maker/checkIfActionTreeLeadsToSetup.ts
