+++ mobile-app/src/utilities/functions/app-maker/getBoxesBasedOnLayout.ts
