+++ mobile-app/src/utilities/functions/app-maker/initialiseNewBox.ts
