+++ mobile-app/src/utilities/functions/app-maker/initialiseNewPage.ts
