+++ mobile-app/src/utilities/functions/app-maker/replaceSetupKeyInActionTree.ts
