+++ mobile-app/src/utilities/functions/app-maker/traverseActionTree.ts
