+++ mobile-app/src/utilities/functions/object-convert/convertStateWithTimestamps.ts
